Scheduler: Fix appointment popup error check test.
